model/pyproject.toml
[project]
name = "model"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "keras-tuner>=1.4.8",
    "matplotlib>=3.10.8",
    "opencv-python>=4.13.0.92",
    "pillow>=12.1.1",
    "pillow-heif>=1.3.0",
    "scikit-learn>=1.8.0",
    "scipy>=1.17.1",
    "seaborn>=0.13.2",
    "tensorboard>=2.20.0",
    "tensorflow[and-cuda]>=2.21.0",
]
